+++ demo/js/modernizr.js
@@ -1,3 +1,15 @@
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
+/*! formstone v1.1.4 [modernizr.js] 2016-06-08 | GPL-3.0 License | formstone.it */
+
 /*! formstone v1.1.4 [modernizr.js] 2016-06-07 | GPL-3.0 License | formstone.it */
 
 /*! formstone v1.1.4 [modernizr.js] 2016-06-07 | GPL-3.0 License | formstone.it */