+++ src/js/lightbox.js
@@ -1,5 +1,7 @@
 /* global define */
 
+// TODO: Roles and labels
+
 (function(factory) {
 	if (typeof define === "function" && define.amd) {
 		define([