+++ src/js/scrollbar.js
@@ -1,5 +1,7 @@
 /* global define */
 
+// TODO: Keyboard controls?
+
 (function(factory) {
 	if (typeof define === "function" && define.amd) {
 		define([